+++ CNC_Sender_GRBL/ENT_Sender_GRBL/Point.cs
@@ -13,15 +13,15 @@ namespace ENT_Sender_GRBL
 
         }
 
-        public Point(decimal x, decimal y, decimal z)
+        public Point(double x, double y, double z)
         {
             X = x;
             Y = y;
             Z = z;
         }
 
-        public decimal X { get; set; }
-        public decimal Y { get; set; }
-        public decimal Z { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
+        public double Z { get; set; }
     }
 }
